requirements.txt
openai==0.26.1
langchain==0.0.65
python-dotenv==0.21.0
faiss-cpu==1.7.3
discord.py
